Update DatabaseInfo.py
Fix the path problem. When runing a file which is not under CNNArt but in a folder, the content of folder 'config' cannot find. By adding the *args and give the file's path, let the .csv file can be read. If run the file directly under the CNNArt, there's no changes, only for running the file under a folder, should invoke this like DatabaseInfo(cfg['MRdatabase'],cfg['subdirs'],folderPath).